image_undersampler/src/undersampler.py
from ultralytics import YOLO
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
import torch
import cv2
from PIL import Image
import os
from typing import List, Dict, Optional, Tuple
from tqdm import tqdm
import gc
from .visualizer import Visualizer
from .statistics import StatisticsGenerator
from .utils import setup_logging, memory_stats
import time

class SimilarityBasedUndersampler:
    def __init__(self, config: Dict):
        """Initialize the undersampler with configuration"""
        self.config = config
        self.logger = setup_logging(level=config['output']['log_level'])
        self.visualizer = Visualizer(config)
        self.statistics = StatisticsGenerator(config)
        
        self._setup_model()
    
    def _setup_model(self):
        """Setup YOLO model"""
        try:
            self.model = YOLO(self.config['model']['path'])
            self.model.fuse()
            self.model.eval()
            self.model.to(self.config['model']['device'])
        except Exception as e:
            self.logger.error(f"Error loading model: {str(e)}")
            raise
    
    # ... [Previous methods for image loading and processing]
    # (I can provide the complete implementation if needed)
    
    def undersample_dataset(self, input_dir: str, output_dir: str) -> Dict:
        """Main method to undersample the dataset"""
        start_time = time.time()
        
        # Setup output directories
        os.makedirs(output_dir, exist_ok=True)
        self.visualizer.set_output_dir(output_dir)
        
        # Process dataset
        # ... [Implementation of the main undersampling logic]
        # (I can provide the complete implementation if needed)
        
        return stats