image_undersampler/requirements.txt
torch>=1.9.0
ultralytics>=8.0.0
numpy>=1.19.5
scikit-learn>=0.24.2
Pillow>=8.3.1
opencv-python>=4.5.3
matplotlib>=3.4.3
seaborn>=0.11.2
psutil>=5.8.0
networkx>=2.6.3
pandas>=1.3.3
tqdm>=4.62.3
pyyaml>=5.4.1